+++ app/views/cart_items/_item.html.erb
@@ -8,7 +8,7 @@
   </div>
   <div class="col m8 s12 left-align">
     <h5>
-      <%= item.title %> by <%= item.user.first_name %> <%= item.user.last_name %>
+      <%= item.title %> by <%= item.vendor.name %>
     </h5>
 
     <p><%= item.description %></p>
